touched up some methods based on others that worked. started on the
newsletter service unit test and will continue once its fully functional

+++ York engineering website/src/main/java/yorkEngineeringSociety/services/NewsletterService.java
@@ -8,7 +8,7 @@ public interface NewsletterService {
 
 	Newsletter findByName(String arg0);
 	
-	Newsletter findByID(int arg0);
+	Newsletter findByID(long arg0);
 	
 	//to change a newsletter later 
 	void changeNewsletterTemplate(Newsletter n, String template);
